backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
asyncpg==0.31.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
Flask==3.1.3
greenlet==3.3.2
h11==0.16.0
hiredis==3.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-anthropic==1.4.0
langchain-core==1.2.26
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
psycopg[binary]==3.3.3
langsmith==0.7.25
Mako==1.3.10
MarkupSafe==3.0.3
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.8
xxhash==3.6.0
zstandard==0.25.0
flask-cors==6.0.2
langchain-google-genai==4.2.1
gunicorn==23.0.0
